+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/ContractNegotiationService.java
@@ -54,14 +54,11 @@ public class ContractNegotiationService {
 
     private final EdcConfiguration config;
 
-    private final CatalogCache catalogCache;
-
     private final PolicyCheckerService policyCheckerService;
 
-    public NegotiationResponse negotiate(final String providerConnectorUrl, final String target)
+    public NegotiationResponse negotiate(final String providerConnectorUrl, final CatalogItem catalogItem)
             throws ContractNegotiationException, UsagePolicyException {
 
-        final CatalogItem catalogItem = catalogCache.getCatalogItem(providerConnectorUrl, target).orElseThrow();
         if (!policyCheckerService.isValid(catalogItem.getPolicy())) {
             log.info("Policy was not allowed, canceling negotiation.");
             throw new UsagePolicyException(catalogItem.getItemId());
@@ -69,7 +66,7 @@ public class ContractNegotiationService {
         final ContractOfferRequest contractOfferRequest = ContractOfferRequest.builder()
                                                                               .offerId(
                                                                                       catalogItem.getItemId())
-                                                                              .assetId(target)
+                                                                              .assetId(catalogItem.getAssetPropId())
                                                                               .policy(catalogItem.getPolicy())
                                                                               .build();
 
@@ -99,7 +96,7 @@ public class ContractNegotiationService {
                                                   .connectorAddress(
                                                           providerConnectorUrl + config.getControlplane().getProviderSuffix())
                                                   .contractId(response.getContractAgreementId())
-                                                  .assetId(target)
+                                                  .assetId(catalogItem.getAssetPropId())
                                                   .dataDestination(destination)
                                                   .build();